requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
djangorestframework==3.13.1
django-filter==21.1
dj-database-url==0.5.0
drf-yasg==1.20.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
firebase_admin==5.2.0